requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
plotly==5.15.0
psycopg2-binary==2.9.6
requests==2.31.0
tenacity==8.2.2
urllib3==2.0.3
Werkzeug==2.3.6
